refactor: app home code split and cleanup
