+++ frontend/utils/meta.ts
@@ -1,0 +1,3 @@
+export function formatTitle(title: string): string {
+  return `${title} | Phase Console`;
+}
